Add custom relations between tables.
The relations are identical to what DBIx would have generated
if the foreign key constraints existed between the id_run
columns of the iseq_run_info and iseq_run tables. The foreign
key is not set in the schema since the order the data appear in
the tables cannot be enforced.