Storage for run params (#179)
* Reformatted and extended the existing test

* Added storage for RunParameters.xml file contents.

Added a table for storing one of Illumina's XML files with
run information, iseq_run_info.

Added collumns to the iseq_run table to capture some values from
these Illumina files. The choice of the values is driven by the
current financial reporting need.

* Add custom relations between tables.

The relations are identical to what DBIx would have generated
if the foreign key constraints existed between the id_run
columns of the iseq_run_info and iseq_run tables. The foreign
key is not set in the schema since the order the data appear in
the tables cannot be enforced.

* Added a method for loading run parameters data.

A new method update_values_from_xml parses XML strings and
loads relevant values to the columns of teh iseq_run table.

* Added a trigger for populating run param values.

IseqRunInfo class: the parent's method is extended to parse the
contents of the run_parameters_xml column and populate relevant
columns in the iseq_run table.

* Set up a proper foreign key. Create parent row.

* extend the workflow_type column to accommodate Standard S4 flowcell

* bring DBIx classes up-to-date